Rename MemberProfiles::PasswordsController to Account::PasswordsController.
