+++ app/controllers/account/base_controller.rb
@@ -1,8 +1,5 @@
 module Account
   class BaseController < ApplicationController
-
     before_action :authenticate_user!
-
-    layout "admin"
   end
 end
