+++ app/views/account/members/show.html.erb
@@ -2,7 +2,7 @@
   <h2 class="new_title">Account Settings</h2>
   <div class="d-flex justify-content-end">
     <button class="secondary-btn account-settings-btn">
-      <%= link_to "Edit Password", edit_member_profiles_password_url %>
+      <%= link_to "Edit Password", edit_account_password_url %>
     </button>
     <br> <br>
     <button class="primary-btn btn-margin">
