+++ app/views/admin/appointments/index.html.erb
@@ -40,16 +40,16 @@
           <% @appointments.each do |appointment| %>
             <tr>
               <td><%= l appointment.starts_at, format: :hour %> - <%= l appointment.ends_at, format: :hour %></td>
-              <td><%= link_to appointment.member.full_name, account_member_path(appointment.member) %></td>
+              <td><%= link_to appointment.member.full_name, admin_member_path(appointment.member) %></td>
               <td><%= appointment.member.pronouns.first %></td>
               <td>
                 <% appointment.holds.each do |hold| %>
                     pick-up
-                    <%= link_to hold.item.name, hold.item %> <br/>
+                    <%= link_to hold.item.name, admin_item_path(hold.item) %> <br/>
                 <% end %>
                 <% appointment.loans.each do |loan| %>
                     drop-off
-                    <%= link_to loan.item.name, loan.item %> <br/>
+                    <%= link_to loan.item.name, admin_item_path(loan.item) %> <br/>
                 <% end %>
               </td>
               <td><%= appointment.comment %></td>
